Fix up some flow control
