+++ pkg/lex/lex.go
@@ -2,147 +2,239 @@ package lex
 
 import (
 	"fmt"
-	"strings"
-	"unicode/utf8"
+	"log"
+	"unicode"
 )
 
-// stateFunction represents the state of the scanner as a function that returns the next state.
-type stateFunction func(*lexerState) stateFunction
+const (
+	substituteHexElement = '\U0000001A'
+	singleQuoteElement   = rune(39) // nolint:gomnd // rune of '
+	alphaNumericElements = "ABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890"
+)
 
-// lexerState holds the state of the scanner.
-type lexerState struct {
-	name      string  // the name of the input; used only for error reports
-	input     string  // the string being scanned
-	pos       Pos     // current position in the input
-	start     Pos     // start position of this item
-	width     Pos     // width of last rune read from input
-	items     []token // channel of scanned items
-	line      int     // 1+number of newlines seen
-	startLine int     // start line of this item
-}
-
-// Lexer is an interface that represents a lexer. It has methods for getting the next token and the name of the lexer.
-type Lexer interface {
-	getNext() token
-	getName() string
-}
-
-// New creates a new scanner for the input string.
-// It takes a name and an input string, and returns a Lexer.
-func New(name, input string) Lexer {
-	lexer := &lexerState{
-		name:      name,
-		input:     input,
-		items:     make([]token, 0),
-		line:      1,
-		startLine: 1,
+// lexStatementTokens lexes tokens within a statement.
+func lexStatementTokens(lexer *lexerState) stateFunction { // nolint:gocyclo // good luck simplifying this
+	switch currentRune := lexer.next(); {
+	case isEOL(currentRune):
+		lexer.emit(tokenEOL)
+	case currentRune == eof:
+		lexer.emit(tokenEOF)
+		return nil
+	case isSpace(currentRune):
+		lexer.backup()
+		return lexSpace
+	case currentRune == picLeft:
+		// special look-ahead for "PIC" so we don't break lexerState.backup().
+		if lexer.pos < Pos(len(lexer.input)) {
+			// Look for PIC
+			if (currentRune < '0' || '9' < currentRune) && lexer.peek() == 'I' && lexer.lookAhead(2) == 'C' { // nolint:gomnd // obvious meaning
+				return lexPICToken
+			}
+			return lexIdentifier
+		}
+	case currentRune == 'O':
+		return lexOccursToken(lexer)
+	case currentRune == 'R':
+		return lexRedefinesToken(lexer)
+	case currentRune == '+' || currentRune == '-' || ('0' <= currentRune && currentRune <= '9'):
+		lexer.backup()
+		return lexNumber
+	case isAlphaNumeric(currentRune):
+		lexer.backup()
+		return lexIdentifier
+	case currentRune == '.':
+		lexer.emit(tokenDot)
+	case currentRune == singleQuoteElement:
+		return lexEnum
+	case currentRune <= unicode.MaxASCII && unicode.IsPrint(currentRune):
+		lexer.emit(tokenChar)
+	case currentRune == substituteHexElement:
+		log.Printf("found SUBSTITUTE rune")
+		lexer.emit(tokenEOF)
+	default:
+		errorMessage := fmt.Errorf("unrecognized character in action: %#U", currentRune)
+		log.Println(errorMessage)
+		return lexer.errorf(errorMessage.Error())
 	}
-	lexer.run()
-	return lexer
+	return lexStatementTokens(lexer)
 }
 
-func (lexer *lexerState) getName() string {
-	return lexer.name
-}
+// lexPICToken lexes a PIC token.
+//
+// TODO: revise this function, flow control is a mess.
+func lexPICToken(lexerState *lexerState) stateFunction {
+	var currentRune rune
+	for {
+		currentRune = lexerState.next()
+		if !isPICChar(currentRune) {
+			// if after a pic character, we get a space it is likely
+			// there may be an OCCURS definition to follow, e.g.
+			// PIC X(10) OCCURS 12.
+			if isSpace(currentRune) {
+				switch nextRune := lexerState.peek(); {
+				case isPICChar(nextRune), isPICType(nextRune), nextRune == '.':
+					continue
 
-// run runs the state machine for the lexer.
-func (lexer *lexerState) run() {
-	for state := lexStatementTokens(lexer); state != nil; {
-		state = state(lexer)
+				default:
+					lexerState.backup()
+					lexerState.emit(tokenPIC)
+					return lexSpace(lexerState)
+				}
+			}
+			// if we've just reached the terminator '.'
+			// let's peek and see if it's actually an explicit decimal point
+			// e.g. PIC 9(9).9(2) -> 123456789.50
+			if isPICChar(lexerState.peek()) {
+				continue
+			}
+			if !lexerState.atPICTerminator() {
+				lexerState.backup()
+				break
+			}
+		}
 	}
+
+	if !lexerState.atPICTerminator() {
+		errorMessage := fmt.Errorf("bad character %#U", currentRune)
+		log.Println(errorMessage)
+		return lexerState.errorf(errorMessage.Error())
+	}
+	lexerState.emit(tokenPIC)
+	return lexStatementTokens(lexerState)
 }
 
-// next returns the next rune in the input.
-func (lexer *lexerState) next() rune {
-	if int(lexer.pos) >= len(lexer.input) {
-		lexer.width = 0
-		return eof
+// lexRedefinesToken is a state function for the lexer. It scans the input for a
+// REDEFINES token. If it encounters a REDEFINES token, it emits a tokenREDEFINES.
+func lexRedefinesToken(lexer *lexerState) stateFunction {
+	if lexer.scanRedefinesToken() {
+		lexer.emit(tokenREDEFINES)
 	}
-	r, width := utf8.DecodeRuneInString(lexer.input[lexer.pos:])
-	lexer.width = Pos(width)
-	lexer.pos += lexer.width
-	if r == '\n' {
-		lexer.line++
-	}
-	return r
+	return lexStatementTokens(lexer)
 }
 
-// peek returns but does not consume the next rune in the input.
-func (lexer *lexerState) peek() rune {
-	r := lexer.next()
+// lexOccursToken is a state function for the lexer. It scans the input for an
+// OCCURS token. If it encounters an OCCURS token, it emits a tokenOCCURS. If
+// it encounters an error while scanning, it returns an error.
+func lexOccursToken(lexer *lexerState) stateFunction {
+	valid, err := lexer.scanOccursToken()
+	if err != nil {
+		return lexer.errorf(err.Error())
+	}
+	if valid {
+		lexer.emit(tokenOCCURS)
+	}
+	return lexStatementTokens(lexer)
+}
+
+// lexIdentifier is a state function for the lexer. It scans the input for alphanumeric
+// characters and emits a tokenBool when it encounters the words "true" or "false",
+// and emits a tokenIdentifier for other words. If it encounters a character that is
+// not a terminator when finished globbing alphanumeric characters, it returns an error.
+func lexIdentifier(lexer *lexerState) stateFunction {
+	for isAlphaNumeric(lexer.next()) {
+	}
 	lexer.backup()
-	return r
-}
 
-// lookAhead returns the rune at the specified position ahead in the input, without consuming any runes.
-func (lexer *lexerState) lookAhead(i int) rune {
-	if int(lexer.pos) >= len(lexer.input) {
-		lexer.width = 0
-		return eof
+	// When finished globbing alphanumeric characters, evaluate the word.
+	word := lexer.input[lexer.start:lexer.pos]
+	if !lexer.atTerminator() {
+		err := fmt.Errorf("bad character %#U", lexer.current())
+		log.Println(err)
+		return lexer.errorf(err.Error())
 	}
-	r, width := utf8.DecodeRuneInString(lexer.input[lexer.pos+Pos(i-1):])
-	lexer.width = Pos(width)
-	lexer.pos += lexer.width
-	if r == '\n' {
-		lexer.line++
+
+	toEmit := tokenIdentifier
+	if word == "true" || word == "false" {
+		toEmit = tokenBool
 	}
-	lexer.backup()
-	return r
+	lexer.emit(toEmit)
+	return lexStatementTokens(lexer)
 }
 
-// backup steps back one rune. Can only be called once per call of next.
-func (lexer *lexerState) backup() {
-	lexer.pos -= lexer.width
-	// Correct newline count.
-	if lexer.width == 1 && lexer.input[lexer.pos] == '\n' {
-		lexer.line--
+// lexEnum is a state function for the lexer. It scans the input for alphanumeric
+// characters and emits a tokenEnum when it encounters a sequence of alphanumeric
+// characters followed by an enum terminator. If it encounters a character that is
+// not alphanumeric and not an enum terminator, it returns an error.
+func lexEnum(lexer *lexerState) stateFunction {
+	for isAlphaNumeric(lexer.peek()) {
+		lexer.acceptRun(alphaNumericElements)
 	}
-}
 
-// emit passes an item back to the client. It takes a token type as a parameter.
-func (lexer *lexerState) emit(tokenType tokenType) {
-	lexer.items = append(lexer.items, token{tokenType, lexer.start, lexer.input[lexer.start:lexer.pos], lexer.startLine})
-	lexer.start = lexer.pos
-	lexer.startLine = lexer.line
-}
-
-// accept consumes the next rune if it's from the valid set of runes.
-// It takes a string of valid runes as a parameter and returns a boolean indicating whether the next rune was accepted.
-func (lexer *lexerState) accept(validRunes string) bool {
-	if strings.ContainsRune(validRunes, lexer.next()) {
-		return true
+	if !lexer.atEnumTerminator() {
+		err := fmt.Errorf("bad character %#U", lexer.current())
+		log.Println(err)
+		return lexer.errorf(err.Error())
 	}
-	lexer.backup()
-	return false
+
+	lexer.next()
+	lexer.emit(tokenEnum)
+	return lexStatementTokens(lexer)
 }
 
-// acceptRun consumes a run of runes from the valid set of runes.
-// It takes a string of valid runes as a parameter.
-func (lexer *lexerState) acceptRun(validRunes string) {
-	for strings.ContainsRune(validRunes, lexer.next()) {
+// lexNumber is a state function for the lexer. It scans the input for numbers
+// and emits a tokenNumber when it encounters a number. If it encounters a complex
+// number (a number followed by '+' or '-' and another number ending with 'i'),
+// it emits a tokenComplex. If the number syntax is incorrect, it returns an error.
+//
+// This isn't a perfect number scanner - for instance it accepts "." and "0x0.2"
+// and "089" - but when it's wrong the input is invalid and the parser (via
+// strconv) will notice.
+func lexNumber(lexer *lexerState) stateFunction {
+	errMsg := "bad number syntax: %q"
+	if !lexer.scanNumber() {
+		return lexer.errorf(errMsg, lexer.input[lexer.start:lexer.pos])
 	}
-	lexer.backup()
-}
-
-// errorf returns an error token and terminates the scan by passing
-// back a nil pointer that will be the next state, terminating lexer.getNext.
-// It takes a string format for the error message and a variadic number of arguments to format the error message.
-func (lexer *lexerState) errorf(errorMessageFormat string, formatArgs ...interface{}) stateFunction {
-	lexer.items = append(lexer.items, token{
-		typ:  tokenError,
-		pos:  lexer.start,
-		val:  fmt.Sprintf(errorMessageFormat, formatArgs...),
-		line: lexer.startLine})
-	return nil
-}
-
-// getNext returns the next item from the input.
-// Called by the parser, not in the lexing goroutine.
-func (lexer *lexerState) getNext() token {
-	if len(lexer.items) == 0 {
-		return token{}
+	toEmit := tokenNumber
+	if sign := lexer.peek(); sign == '+' || sign == '-' {
+		// Complex: 1+2i. No spaces, must end in 'i'.
+		if !lexer.scanNumber() || lexer.input[lexer.pos-1] != 'i' {
+			return lexer.errorf(errMsg, lexer.input[lexer.start:lexer.pos])
+		}
+		toEmit = tokenComplex
 	}
-	var next token
-	next, lexer.items = lexer.items[0], lexer.items[1:]
-	return next
+	lexer.emit(toEmit)
+	return lexStatementTokens(lexer)
+}
+
+// lexSpace is a state function for the lexer. It scans the input for spaces
+// and emits a tokenSpace when it encounters one or more spaces. After emitting,
+// it transitions the lexer to the lexStatementTokens state.
+func lexSpace(lexer *lexerState) stateFunction {
+	for isSpace(lexer.peek()) {
+		lexer.next()
+	}
+	lexer.emit(tokenSpace)
+	return lexStatementTokens(lexer)
+}
+
+// isSpace reports whether the currentChar is a space character.
+// It checks if the currentChar is either a space (' ') or a tab ('\t').
+func isSpace(currentChar rune) bool {
+	return currentChar == ' ' || currentChar == '\t'
+}
+
+// isEOL reports whether the currentChar is an end-of-line character.
+// It checks if the currentChar is either a carriage return ('\r') or a newline ('\n').
+func isEOL(currentChar rune) bool {
+	return currentChar == '\r' || currentChar == '\n'
+}
+
+// isAlphaNumeric reports whether the currentChar is an alphabetic, digit, or underscore.
+// It checks if the currentChar is an underscore ('_'), a hyphen ('-'), a letter, or a digit.
+func isAlphaNumeric(currentChar rune) bool {
+	return currentChar == '_' || currentChar == '-' || unicode.IsLetter(currentChar) || unicode.IsDigit(currentChar)
+}
+
+// isPICChar reports whether the currentChar is a PIC character.
+// It checks if the currentChar is a PIC character or a number.
+func isPICChar(currentChar rune) bool {
+	_, ok := picChars[currentChar]
+	return ok || unicode.IsNumber(currentChar)
+}
+
+// isPICType reports whether the currentChar is a PIC type.
+// It checks if the currentChar is a PIC type.
+func isPICType(currentChar rune) bool {
+	_, ok := picTypes[currentChar]
+	return ok
 }
